requirements.txt
streamlit
openai
pandas
plotly
numpy
scikit-learn
joblib
matplotlib
dotenv
statsmodels
azure-storage-blob
pyopenssl
streamlit-extras